pyproject.toml
[tool.poetry]
name = "image-server"
version = "0.1.0"
description = ""
authors = ["Alexandra Ispas <aispas@adobe.com>"]
readme = "README.md"
packages = [{include = "image_server", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi-sqlalchemy = "^0.2.1"
psycopg2 = "^2.9.10"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
